requirements-preview.txt
-e git+https://github.com/adafruit/Adafruit_Blinka_Raspberry_Pi5_Piomatter.git@27c2198e894f15813889b16dba400dd3e4311058#egg=Adafruit_Blinka_Raspberry_Pi5_Piomatter
audioread==3.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
decorator==5.2.1
idna==3.11
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
msgpack==1.1.2
numba==0.62.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
pycparser==2.23
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
soundfile==0.13.1
soxr==1.0.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
